🚀 refactor(lotto): 로또 결과 Map을 포장하는 일급 컬렉션 LottoResults 생성
