+++ src/main/java/lotto/domain/LottoGame.java
@@ -1,8 +1,5 @@
 package lotto.domain;
 
-import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -19,36 +16,14 @@ public class LottoGame {
         this.winningLotto = winningLotto;
     }
 
-    public Map<Ranking, Integer> findWinner() {
-
+    public LottoResults findWinner() {
         final Map<Ranking, Integer> totalResult = new HashMap<>();
-        initializeTotalResult(totalResult);
+        final LottoResults lottoResults = new LottoResults(totalResult);
 
         final List<Ranking> rankings = lottos.judgeAllUserLotto(winningLotto);
-        rankings.forEach(ranking -> updateResults(totalResult, ranking));
+        rankings.forEach(ranking -> lottoResults.updateResults(ranking));
 
-        return totalResult;
-    }
-
-    private void initializeTotalResult(final Map<Ranking, Integer> totalResult) {
-        Arrays.stream(Ranking.values())
-            .forEach(ranking -> totalResult.put(ranking, ZERO));
-    }
-
-    private void updateResults(final Map<Ranking, Integer> totalResult, final Ranking ranking) {
-        Integer currentCount = totalResult.get(ranking);
-        totalResult.put(ranking, ++currentCount);
-    }
-
-    public BigDecimal calculateYield(final Map<Ranking, Integer> totalResult, final Money money) {
-        final List<Ranking> rankings = new ArrayList<>(totalResult.keySet());
-
-        final BigDecimal totalWinnerPrize = BigDecimal.valueOf(rankings.stream()
-            .mapToInt(ranking ->
-                ranking.multiplyCountAndWinnerPrice(totalResult.get(ranking)))
-            .sum());
-
-        return totalWinnerPrize.divide(BigDecimal.valueOf(money.getMoney()));
+        return lottoResults;
     }
 
     public Lottos getLottos() {
